@@ -11,15 +11,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.4.0",
-    "@angular/common": "~13.4.0",
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
     "@angular/fire": "^6.0.2",
-    "@angular/compiler": "~13.4.0",
-    "@angular/core": "~13.4.0",
-    "@angular/forms": "~13.4.0",
-    "@angular/platform-browser": "~13.4.0",
-    "@angular/platform-browser-dynamic": "~13.4.0",
-    "@angular/router": "~13.4.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "ngx-typed-js": "^2.1.1",
     "rxjs": "~6.6.0",
     "swiper": "^8.1.5",
@@ -27,14 +27,14 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.11",
+    "@angular-devkit/build-angular": "^14.2.13",
     "@angular-eslint/builder": "12.7.0",
     "@angular-eslint/eslint-plugin": "12.7.0",
     "@angular-eslint/eslint-plugin-template": "12.7.0",
     "@angular-eslint/schematics": "12.7.0",
     "@angular-eslint/template-parser": "12.7.0",
-    "@angular/cli": "~13.3.11",
-    "@angular/compiler-cli": "~13.4.0",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
     "@types/jasmine": "~3.8.0",
     "@types/node": "^12.11.1",
     "@typescript-eslint/eslint-plugin": "4.28.2",
